chore(aws): move build cache related code to own file
